Extract `PushMessageBuilder` and add unit tests; remove Android-instrumented `PushRegistrationManagerTest`
